backend/auth-service/requirements.txt
<<<<<<< HEAD
# backend/auth_service/requirements.txt

fastapi==0.68.0
uvicorn==0.15.0
sqlalchemy==1.4.23
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.5
python-dotenv==0.19.0
pydantic[email]==1.8.2 
=======
fastapi>=0.109.2
uvicorn>=0.27.1
sqlalchemy>=2.0.27
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.9
python-dotenv>=1.0.1
pydantic[email]>=2.6.1
bcrypt==3.2.2 
>>>>>>> 4364be74933165036c85e9fd06094caaf0309666
